최종

+++ team_project1/src/main/java/action/review/RevDeleteProAction.java
@@ -2,6 +2,7 @@ package action.review;
 
 import java.io.File;
 import java.io.PrintWriter;
+import java.util.Enumeration;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
@@ -21,6 +22,16 @@ public class RevDeleteProAction implements Action {
 		int board_no=Integer.parseInt(request.getParameter("board_no"));
 		 String fileName = request.getParameter("rev_fileName");
 		
+		 Enumeration params = request.getParameterNames();
+			System.out.println("----------------------------");
+			while (params.hasMoreElements()){
+			    String name = (String)params.nextElement();
+			    System.out.println(name + " : " +request.getParameter(name));
+			}
+			System.out.println("----------------------------");
+		 
+		 
+		 
 		//글작성자와 로그인작성자가 같으면 수정 삭제버튼이 보이겠지만 get방식으로 홈페이지에 바로접속할 수 있으므로 한번더 걸러주기
 		HttpSession session = request.getSession();
 		String mem_id = (String)session.getAttribute("mem_id");
@@ -28,7 +39,7 @@ public class RevDeleteProAction implements Action {
 		
 		RevDeleteProService revDeleteProService = new RevDeleteProService();
 		boolean isArticleWriter =revDeleteProService.isArticleWriter(board_no, mem_id);
-
+/*
 		if(!isArticleWriter){
 			response.setContentType("text/html;charset=UTF-8");
 			PrintWriter out=response.getWriter();
@@ -40,7 +51,7 @@ public class RevDeleteProAction implements Action {
 		}
 
 		else{
-			
+			*/
 			//삭제작업
 			
 			boolean isDeleteSuccess = revDeleteProService.removeArticle(board_no);
@@ -56,20 +67,21 @@ public class RevDeleteProAction implements Action {
 			}
 			else{//삭제작업 성공시 저장된 파일도 지워야함
 				 
-			        
+			     if(fileName!=null) {
 				  ServletContext context = request.getServletContext();
 					String uploadPath = context.getRealPath("/reviewUpload");
 			        
 			        File file = new File(uploadPath+File.separator +fileName);
-			        File file2 = new File(uploadPath + File.separator + "sm_" + fileName);
+
+			    
 			        
 			        if(file.exists()) {    //삭제하고자 하는 파일이 해당 서버에 존재하면 삭제시킨다..바로 하면 삭제가안됨..?ㅎ
 			        	//Thread.sleep(1000); 
 			        	file.delete();
-			            file2.delete();
+			        	
 			        }
-			    
-				
+			     }
+			     
 				
 				
 				
@@ -77,7 +89,7 @@ public class RevDeleteProAction implements Action {
 
 			}
 			
-		}
+		/*}*/
 
 
 		return forward;
